+++ src/const.rs
@@ -5,9 +5,7 @@
 #![allow(dead_code)]
 
 use lazy_static::lazy_static;
-
-/// The LLVM framework version.
-pub const LLVM_VERSION: semver::Version = semver::Version::new(15, 0, 4);
+use sha3::Digest;
 
 /// The `FREE_VAR_SPACE` offset.
 pub const OFFSET_FREE_VAR_SPACE: usize = 0;
@@ -54,8 +52,9 @@ lazy_static! {
     /// The Vyper forwarder bytecode in bytes.
     ///
     pub static ref FORWARDER_CONTRACT_BYTECODE_WORDS: Vec<[u8; compiler_common::BYTE_LENGTH_FIELD]> = {
+        let metadata_hash = sha3::Keccak256::digest(FORWARDER_CONTRACT_ASSEMBLY.as_bytes()).into();
         let mut assembly =
-            zkevm_assembly::Assembly::try_from(FORWARDER_CONTRACT_ASSEMBLY.to_owned()).expect("Always valid");
+            zkevm_assembly::Assembly::from_string(FORWARDER_CONTRACT_ASSEMBLY.to_owned(), metadata_hash).expect("Always valid");
         assembly
             .compile_to_bytecode().expect("Always valid")
     };