+++ src/project/contract/llvm_ir.rs
@@ -3,6 +3,7 @@
 //!
 
 use crate::build::contract::Contract as ContractBuild;
+use crate::project::contract::metadata::Metadata as ContractMetadata;
 use crate::project::dependency_data::DependencyData;
 
 ///
@@ -10,16 +11,16 @@ use crate::project::dependency_data::DependencyData;
 ///
 #[derive(Debug, Clone)]
 pub struct Contract {
-    /// The source code.
-    pub code: String,
+    /// The contract source code.
+    pub source_code: String,
 }
 
 impl Contract {
     ///
     /// A shortcut constructor.
     ///
-    pub fn new(code: String) -> Self {
-        Self { code }
+    pub fn new(source_code: String) -> Self {
+        Self { source_code }
     }
 
     ///
@@ -28,19 +29,29 @@ impl Contract {
     pub fn compile(
         self,
         contract_path: &str,
+        source_code_hash: [u8; compiler_common::BYTE_LENGTH_FIELD],
         target_machine: compiler_llvm_context::TargetMachine,
         optimizer_settings: compiler_llvm_context::OptimizerSettings,
         debug_config: Option<compiler_llvm_context::DebugConfig>,
     ) -> anyhow::Result<ContractBuild> {
         let llvm = inkwell::context::Context::create();
+        let optimizer = compiler_llvm_context::Optimizer::new(target_machine, optimizer_settings);
+
+        let metadata_hash = ContractMetadata::new(
+            &source_code_hash,
+            &compiler_llvm_context::LLVM_VERSION,
+            semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid"),
+            optimizer.settings().to_owned(),
+        )
+        .keccak256();
+
         let memory_buffer = inkwell::memory_buffer::MemoryBuffer::create_from_memory_range_copy(
-            self.code.as_bytes(),
+            self.source_code.as_bytes(),
             contract_path,
         );
         let module = llvm
             .create_module_from_ir(memory_buffer)
             .map_err(|error| anyhow::anyhow!(error.to_string()))?;
-        let optimizer = compiler_llvm_context::Optimizer::new(target_machine, optimizer_settings);
         let context = compiler_llvm_context::Context::<DependencyData>::new(
             &llvm,
             module,
@@ -48,7 +59,8 @@ impl Contract {
             None,
             debug_config,
         );
-        let build = context.build(contract_path)?;
+
+        let build = context.build(contract_path, metadata_hash)?;
 
         Ok(ContractBuild::new(build))
     }