+++ src/zkvyper/arguments.rs
@@ -31,16 +31,11 @@ pub struct Arguments {
     #[structopt(long = "overwrite")]
     pub overwrite: bool,
 
-    /// Disable the LLVM bytecode optimizer.
-    #[structopt(long = "no-optimize")]
-    pub no_optimize: bool,
+    /// Sets the optimization parameter -O[0 | 1 | 2 | 3 | s | z].
+    #[structopt(short = "O", long = "optimization")]
+    pub optimization: Option<char>,
 
-    /// Sets the LLVM optimizer options.
-    #[structopt(long = "llvm-opt")]
-    pub llvm_options: Option<String>,
-
-    /// Path to the `vyper` executable.
-    /// By default, the one in $PATH is used.
+    /// Path to the `vyper` executable. By default, the one in `${PATH}` is used.
     #[structopt(long = "vyper")]
     pub vyper: Option<String>,